requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
fsspec==2025.9.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.10
scipy==1.16.2
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing_extensions==4.15.0
ultralytics==8.3.204
ultralytics-thop==2.0.17
urllib3==2.5.0
